Change type of TSPAnt#tourLength to double

+++ hw06-0036505665/src/main/java/hr/fer/zemris/optjava/dz6/TSPAnt.java
@@ -13,7 +13,7 @@ public class TSPAnt {
     /**
      * The length of this ant's tour.
      */
-    public int tourLength;
+    public double tourLength;
 
     /**
      * The tour of cities.
